+++ apps/pie-storybook/stories/pie-card.stories.ts
@@ -1,10 +1,11 @@
 import { nothing } from 'lit';
 import { html } from 'lit/static-html.js';
+import { ifDefined } from 'lit/directives/if-defined.js';
 
 /* eslint-disable import/no-duplicates */
 import '@justeattakeaway/pie-card';
 import {
-    CardProps as CardPropsBase, variants, tags, paddingValues,
+    CardProps as CardPropsBase, variants, tags, paddingValues, defaultProps,
 } from '@justeattakeaway/pie-card';
 /* eslint-enable import/no-duplicates */
 
@@ -15,17 +16,14 @@ type CardProps = SlottedComponentProps<CardPropsBase>;
 type CardStoryMeta = StoryMeta<CardProps>;
 
 const defaultArgs: CardProps = {
-    tag: 'button',
-    variant: 'default',
+    ...defaultProps,
     href: '#',
     padding: 'a',
-    disabled: false,
     rel: '',
     target: '',
     aria: {
         label: 'Click to go to restaurant',
     },
-    isDraggable: false,
     // This is just an arbitrary example of some markup a user may pass into the card
     slot: `<div style="font-size: calc(var(--dt-font-body-l-size) * 1px); font-family: var(--dt-font-interactive-l-family);">
         <h2 style="margin-top: 0"> Card title </h2>
@@ -49,7 +47,7 @@ const cardStoryMeta: CardStoryMeta = {
             control: 'select',
             options: tags,
             defaultValue: {
-                summary: 'button',
+                summary: defaultProps.tag,
             },
         },
         variant: {
@@ -57,7 +55,7 @@ const cardStoryMeta: CardStoryMeta = {
             control: 'select',
             options: variants,
             defaultValue: {
-                summary: 'default',
+                summary: defaultProps.variant,
             },
         },
         padding: {
@@ -71,7 +69,7 @@ const cardStoryMeta: CardStoryMeta = {
             description: 'If `true`, disables the card.',
             control: 'boolean',
             defaultValue: {
-                summary: false,
+                summary: defaultProps.disabled,
             },
         },
         href: {
@@ -100,7 +98,7 @@ const cardStoryMeta: CardStoryMeta = {
         isDraggable: {
             control: 'boolean',
             defaultValue: {
-                summary: false,
+                summary: defaultProps.isDraggable,
             },
         },
     },
@@ -128,8 +126,8 @@ const Template: TemplateFunction<CardProps> = ({
     isDraggable,
 }) => html`
         <pie-card
-            tag="${tag}"
-            variant="${variant}"
+            tag="${ifDefined(tag)}"
+            variant="${ifDefined(variant)}"
             href="${href || nothing}"
             target="${target || nothing}"
             rel="${rel || nothing}"