+++ packages/components/pie-link/src/defs.ts
@@ -1,3 +1,5 @@
+import { type ComponentDefaultPropsGeneric } from '@justeattakeaway/pie-webc-core';
+
 export const variants = ['default', 'high-visibility', 'inverse'] as const;
 export const sizes = ['small', 'medium'] as const;
 export const iconPlacements = ['leading', 'trailing'] as const;
@@ -65,3 +67,17 @@ export interface LinkProps {
      */
     type?: typeof buttonTypes[number];
 }
+
+export type DefaultProps = ComponentDefaultPropsGeneric<LinkProps, 'tag' | 'variant' | 'size' | 'underline' | 'iconPlacement' | 'isBold' | 'isStandalone' | 'hasVisited' | 'type'>;
+
+export const defaultProps: DefaultProps = {
+    tag: 'a',
+    variant: 'default',
+    size: 'medium',
+    underline: 'default',
+    iconPlacement: 'leading',
+    isBold: false,
+    isStandalone: false,
+    hasVisited: false,
+    type: 'submit',
+};