+++ packages/components/pie-webc-core/src/index.ts
@@ -1,4 +1,5 @@
 export * from './decorators';
 export * from './functions';
 export * from './interfaces';
+export * from './types';
 export * from './mixins';