+++ packages/components/pie-modal/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @justeattakeaway/pie-modal
 
+## 0.42.6
+
+### Patch Changes
+
+- [Changed] - components to use a single source of truth for props default values ([#1451](https://github.com/justeattakeaway/pie/pull/1451)) by [@fernandofranca](https://github.com/fernandofranca)
+
+- Updated dependencies [[`bfb142f84`](https://github.com/justeattakeaway/pie/commit/bfb142f848a06a8ac6662b300289e34eb22bb6bd), [`bfb142f84`](https://github.com/justeattakeaway/pie/commit/bfb142f848a06a8ac6662b300289e34eb22bb6bd), [`abaadfddc`](https://github.com/justeattakeaway/pie/commit/abaadfddc651f9b2516e94449211bbd19a1bf651)]:
+  - @justeattakeaway/pie-webc-core@0.23.0
+  - @justeattakeaway/pie-icon-button@0.28.6
+  - @justeattakeaway/pie-spinner@0.6.5
+  - @justeattakeaway/pie-button@0.47.5
+  - @justeattakeaway/pie-icons-webc@0.24.0
+
 ## 0.42.5
 
 ### Patch Changes